backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart>=0.0.12
pydantic[email]>=2.10.0
python-dotenv==1.0.1
# Text processing
httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.3.0
pypdf2==3.0.1
python-docx>=1.1.2
openai>=1.51.0
pymongo>=4.6.0
